add zero function

+++ lib/util.py
@@ -1,5 +1,7 @@
-def add(a: int, b: int) -> int:
+def add(a: float, b: float) -> float:
     return a+b
 
-def sub(a: int, b: int) -> int:
+def sub(a: float, b: float) -> float:
     return a-b
+
+
